Conflicts:
	Integration/Integration.sqlproj
	Integration/Scripts/PreDeployment.sql
	i21Database/Scripts/Post-Deployment/SM/DefaultData/1_MasterMenu.sql
	i21Database/Scripts/Post-Deployment/SM/SMDataFixes.sql
	i21Database/dbo/Functions/fnTRSearchItemId.sql
	i21Database/dbo/Stored Procedures/uspARPostInvoice.sql
	i21Database/dbo/Stored Procedures/uspICAddItemShipment.sql
	i21Database/dbo/Stored Procedures/uspICInventoryReceiptAfterSave.sql
	i21Database/dbo/Stored Procedures/uspPRUpdateInsertForm941.sql
	i21Database/dbo/Stored Procedures/uspRKGetAllElectronicPricingURL.sql
	i21Database/dbo/Tables/tblSCTicket.sql
	i21Database/dbo/Tables/tblTMDeliveryHistory.sql
	i21Database/dbo/User Defined Types/ShipmentStagingTable.sql
	i21Database/dbo/Views/vyuARPaymentReport.sql
	i21Database/i21Database.refactorlog
	i21Database/i21Database.sqlproj
	i21PreDeployment/Scripts/Pre-Deployment/20_CF/FixeDataWithContraints.sql

+++ i21Database/dbo/Stored Procedures/uspCCTransactionToAPBill.sql
@@ -19,7 +19,7 @@ DECLARE @ErrorSeverity INT,
 		@ErrorNumber   INT,
 		@ErrorState INT
 
-BEGIN TRY
+BEGIN
 
 	IF(@post = 1)
 		BEGIN
@@ -76,30 +76,39 @@ BEGIN TRY
 			WHERE intAccountId IS NOT NULL
 			GROUP BY  intAccountId, intSiteDetailId, strItem 
 
-			EXEC [dbo].[uspAPCreateBillData]
-				 @userId	= @userId
-				,@vendorId = @vendorId
-				,@type = 3	
-				,@shipTo = @shipTo
-				,@vendorOrderNumber = @ccdReference
-				,@voucherDate = @dtmDate
-				,@voucherDetaiCC = @voucherDetailCC
-				,@billId = @createdBillId OUTPUT
+			DECLARE @intCountDetail INT = 0
 
-			UPDATE tblAPBill SET strComment = @ccdReference WHERE intBillId = @createdBillId
+			SELECT @intCountDetail = COUNT(*) FROM @voucherDetailCC
 
-			EXEC [dbo].[uspAPPostBill]
-				@post = @post
-				,@recap = @recap
-				,@isBatch = 0
-				,@param = @createdBillId
-				,@userId = @userId
-				,@success = @success OUTPUT
+			IF(@intCountDetail > 0)
+			BEGIN
+				EXEC [dbo].[uspAPCreateBillData]
+					 @userId	= @userId
+					,@vendorId = @vendorId
+					,@type = 3	
+					,@shipTo = @shipTo
+					,@vendorOrderNumber = @ccdReference
+					,@voucherDate = @dtmDate
+					,@voucherDetaiCC = @voucherDetailCC
+					,@billId = @createdBillId OUTPUT
+
+				UPDATE tblAPBill SET strComment = @ccdReference WHERE intBillId = @createdBillId
+
+				EXEC [dbo].[uspAPPostBill]
+					@post = @post
+					,@recap = @recap
+					,@isBatch = 0
+					,@param = @createdBillId
+					,@userId = @userId
+					,@success = @success OUTPUT
+			END
+
 		END
 	ELSE IF (@post = 0)
 		BEGIN
 			
 			DECLARE @billId	INT = NULL
+			DECLARE @isPaid BIT = NULL
 			
 			--Find AP Info
 			SELECT @billId = C.intBillId FROM tblCCSiteHeader A 
@@ -108,35 +117,33 @@ BEGIN TRY
 			WHERE A.intSiteHeaderId = @intSiteHeaderId
 			GROUP BY C.intBillId
 
+			SELECT @isPaid = ysnPaid FROM tblAPBill WHERE intBillId = @billId
+
 			IF(@billId IS NOT NULL)
 			BEGIN
-				EXEC [dbo].[uspAPPostBill]
-					@post = @post
-					,@recap = 0
-					,@isBatch = 0
-					,@param = @billId
-					,@userId = @userId
-					,@success = @success OUTPUT
 
-				--DELETE Bill Transaction
-				DELETE FROM tblAPBill WHERE intBillId = 
-				(SELECT DISTINCT intBillId 
-					FROM tblAPBillDetail A
-				JOIN  tblCCSiteDetail B ON A.intCCSiteDetailId = B.intSiteDetailId
-					WHERE B.intSiteHeaderId = @intSiteHeaderId)
+				IF(@isPaid = 1)
+					BEGIN
+						RAISERROR('Cannot unpost this transaction. There is already payment made on the associated Voucher/Invoice.', 16, 1)
+					END
+				ELSE
+					BEGIN
+						EXEC [dbo].[uspAPPostBill]
+							@post = @post
+							,@recap = 0
+							,@isBatch = 0
+							,@param = @billId
+							,@userId = @userId
+							,@success = @success OUTPUT
 
+						--DELETE Bill Transaction
+						DELETE FROM tblAPBill WHERE intBillId = 
+						(SELECT DISTINCT intBillId 
+							FROM tblAPBillDetail A
+						JOIN  tblCCSiteDetail B ON A.intCCSiteDetailId = B.intSiteDetailId
+							WHERE B.intSiteHeaderId = @intSiteHeaderId)
+					END
 			END
-			ELSE
-				RAISERROR('Bill ID is null', 0, 1)
-	
-		END
 
-END TRY
-BEGIN CATCH
-	SET @ErrorSeverity = ERROR_SEVERITY()
-	SET @ErrorNumber   = ERROR_NUMBER()
-	SET @errorMessage  = ERROR_MESSAGE()
-	SET @ErrorState    = ERROR_STATE()
-	SET	@success = 0
-	RAISERROR (@errorMessage , @ErrorSeverity, @ErrorState, @ErrorNumber)
-END CATCH
+		END
+END
